+++ sdk/storage/azure-storage-file-datalake/src/samples/java/com/azure/storage/file/datalake/FileSystemAsyncClientJavaDocCodeSamples.java
@@ -223,7 +223,7 @@ public class FileSystemAsyncClientJavaDocCodeSamples {
         String owner = "rwx";
         String group = "r--";
         String leaseId = UUID.randomUUID().toString();
-        int duration = 15;
+        Long duration = 15L;
         DataLakeAccessOptions accessOptions = new DataLakeAccessOptions()
             .setPermissions(permissions)
             .setUmask(umask)
@@ -305,7 +305,7 @@ public class FileSystemAsyncClientJavaDocCodeSamples {
         String owner = "rwx";
         String group = "r--";
         String leaseId = UUID.randomUUID().toString();
-        int duration = 15;
+        Long duration = 15L;
         DataLakeAccessOptions accessOptions = new DataLakeAccessOptions()
             .setPermissions(permissions)
             .setUmask(umask)
