+++ sdk/storage/azure-storage-file-datalake/src/samples/java/com/azure/storage/file/datalake/PathAsyncClientJavaDocCodeSamples.java
@@ -88,7 +88,7 @@ public class PathAsyncClientJavaDocCodeSamples {
         String owner = "rwx";
         String group = "r--";
         String leaseId = UUID.randomUUID().toString();
-        int duration = 15;
+        Long duration = 15L;
         DataLakeAccessOptions accessOptions = new DataLakeAccessOptions()
             .setPermissions(permissions)
             .setUmask(umask)
