+++ sdk/storage/azure-storage-file-datalake/src/samples/java/com/azure/storage/file/datalake/PathClientJavaDocCodeSamples.java
@@ -98,7 +98,7 @@ public class PathClientJavaDocCodeSamples {
         String owner = "rwx";
         String group = "r--";
         String leaseId = UUID.randomUUID().toString();
-        int duration = 15;
+        Long duration = 15L;
         DataLakeAccessOptions accessOptions = new DataLakeAccessOptions()
             .setPermissions(permissions)
             .setUmask(umask)
